Iterate rather than enumate
Previously, get_car_links would enumerate car_list.

Enumeration is the easy Pythonic way do code for-each loops while
still having access to the index of the element you're currently
on.

However, the code following the enumeration made no use of the
index which was provided to local variable i.

Accordingly, delete the unused variable i, and change the
enumeration to a mere Pythonic iteration.