Use .py file to define functions
Previously, the functions being called in this notebook were
defined in the notebook itself.

However, Issue #4 prefers to define the functions in use for
this notebook in a standalone .py file.

Accordingly, having copied those functions into a standalone
.py file, simply import that file, prepend function calls
with cr., and move unneeded imports out of the Notebook.

This commit is intended to fix #4.