Make get_city_from_url more robust
Previously, get_city_from_url was Pythonic, which is great.

Unfortunately, this didn't allow for error checking, such as the string
being passed is too short, or doesn't contain a city, etc.

Accordingly, revise the function to be more C++ish, iterating through
the passed string until the first `.` is located, which is the index
that postpends the city.

Since the url being passed should always begin with `http://`, we can
further assume that the city name begins at the index after that string,
namely index `7`.

Note, however, that if the host website decides to use https://, further
logic will be required to return a valid city name.
Or, put differently, if the host website ever uses `https://`, expect to
see a lot of `/city` entries.