Fix comparison compatibility
Previously, the list was cast to a string, which was then compared
against a string representation of a list.
However, this is not necessary, as Python allows you to compare
lists to each other.
Additionally, although this approach worked in Python2, it did
not work in Python3.

Accordingly, compare lists rather than strings.